+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+- Support for non-streaming downloads and the corresponding `--no-stream` flag to the CLI ([#208](https://github.com/stac-utils/stac-asset/pull/208))
+
 ## [0.4.2] - 2024-07-28
 
 ### Added